+++ src/components/index.ts
@@ -80,6 +80,7 @@ export { default as TabBar } from './Navigation/TabBar'
 // -- Overlays -- //
 
 export { default as Menu } from './Overlays/Menu'
+export { default as Modal } from './Overlays/Modal'
 
 // -- Overlays -- //